+++ cmake/ExternLibs.cmake
@@ -14,7 +14,7 @@ function(FETCHSHAREDTYPES)
 	FetchContent_Declare(
 	  shared_types
 	  GIT_REPOSITORY https://github.com/Anticarium/Shared_Types.git
-	  GIT_TAG        v1.1.3
+          GIT_TAG        784ee41d2e200b53f70c494bffa313822811ba9d
 	)
 
 	FetchContent_GetProperties(shared_types)
@@ -33,4 +33,15 @@ function(ADDSHAREDTYPES TARGET_NAME)
 			$<INSTALL_INTERFACE:include>)
 
 	TARGET_LINK_LIBRARIES(${TARGET_NAME} PUBLIC shared_types_include shared_types_lib)
+endfunction()
+
+function(ADDGTEST TARGET_NAME)
+        include(FetchContent)
+        FetchContent_Declare(
+                googletest
+                URL https://github.com/google/googletest/archive/2f80c2ba71c0e8922a03b9b855e5b019ad1f7064.zip
+        )
+        set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
+        target_link_libraries(${TARGET_NAME} PRIVATE gtest_main)
+        FetchContent_MakeAvailable(googletest)
 endfunction()