+++ src/anticarium_pi/TestClass.cpp
@@ -1,6 +1,0 @@
-#include "TestClass.h"
-
-TestClass::TestClass()
-{
-
-}